@@ -12,11 +12,11 @@
     "analyze": "webpack --config webpack.analyze.ts"
   },
   "dependencies": {
-    "@emotion/react": "^11.4.0",
-    "@emotion/styled": "^11.3.0",
+    "@emotion/react": "^11.6.0",
+    "@emotion/styled": "^11.6.0",
     "@fontsource/lato": "^4.5.0",
     "@loadable/component": "^5.15.0",
-    "@mui/material": "^5.0.6",
+    "@mui/material": "^5.1.1",
     "@perses-ui/core": "^1.0.0",
     "@perses-ui/panels-plugin": "^1.0.0",
     "@perses-ui/prometheus-plugin": "^1.0.0",
@@ -26,27 +26,27 @@
     "notistack": "^2.0.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-error-boundary": "^3.1.3",
-    "react-query": "^3.31.0",
+    "react-error-boundary": "^3.1.4",
+    "react-query": "^3.32.3",
     "use-immer": "^0.6.0",
-    "use-resize-observer": "^7.0.1"
+    "use-resize-observer": "^8.0.0"
   },
   "devDependencies": {
     "@svgr/webpack": "^5.5.0",
     "@types/loadable__component": "^5.13.4",
     "@types/webpack-bundle-analyzer": "^4.4.1",
-    "@types/webpack-dev-server": "^4.3.0",
-    "css-loader": "^6.5.0",
+    "@types/webpack-dev-server": "^4.3.1",
+    "css-loader": "^6.5.1",
     "dotenv-defaults": "^3.0.0",
     "fork-ts-checker-webpack-plugin": "^6.4.0",
     "html-webpack-plugin": "^5.5.0",
     "style-loader": "^3.3.1",
     "ts-loader": "^9.2.6",
     "ts-node": "^10.4.0",
-    "webpack": "^5.61.0",
+    "webpack": "^5.64.1",
     "webpack-bundle-analyzer": "^4.5.0",
     "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^4.4.0",
+    "webpack-dev-server": "^4.5.0",
     "webpack-merge": "^5.8.0"
   },
   "jest": {
